requirements.txt
flask
jinja2
sqlalchemy
pyyaml
